% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine}
\alias{combine}
\title{Combine Ensembles of Trees}
\usage{
combine(...)
}
\arguments{
\item{..., }{two or more objects of class \code{rfCountData}, to be combined into one.}
}
\value{
An object of class \code{rfCountData}.
}
\description{
Combine two more more ensembles of trees into one.
}
\details{
combine
}
\note{
The confusion, \code{err.rate}, \code{mse} and \code{rsq} components
 (as well as the corresponding components in the test compnent, if exist) of the combined object will be \code{NULL}.
}
\author{
Andy Liaw \email{andy_liaw@merck.com}
}
