% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTree.R
\name{getTree}
\alias{getTree}
\title{getTree}
\usage{
getTree(rfobj, k=1, labelVar=FALSE)
}
\arguments{
\item{rfobj, }{a \code{rfCountData} object.}

\item{k, }{which tree to extract ?}

\item{labelVar, }{Should better labels be used for splitting variables and predicted class ?}
}
\value{
A matrix (or data frame, if \code{labelVar=TRUE}) with six columns 
and number of rows equal to total number of nodes in the tree.  The six columns are: \cr
\code{left daughter}, the row where the left daughter node is; 0 if the node is terminal.\cr
\code{right daughter}, the row where the right daughter node is; 0 if the node is terminal.\cr
\code{split var}, which variable was used to split the node; 0 if the  node is terminal.\cr
\code{split point}, where the best split is; see Details for  categorical predictor.\cr
\code{status}, is the node terminal (-1) or not (1).\cr
\code{prediction},the prediction for the node; 0 if the node is not terminal.
}
\description{
This function extract the structure of a tree from a \code{rfCountData} object.
}
\details{
For numerical predictors, data with values of the variable less than or equal to the 
splitting point go to the left daughter node. \cr
For categorical predictors, the splitting point is represented by an integer, whose binary 
expansion gives the identities of the categories that goes to left or right. \cr
For example, if a predictor has four categories, and the split point is 13. 
The binary expansion of 13 is (1, 0, 1, 1) (because $13 = 1*2^0 + 0*2^1$ + $1*2^2 + 1*2^3$),
so cases with categories 1, 3, or 4 in this predictor get sent to the left, and the rest to the right.
}
\seealso{
\link{rfPoisson}
}
\author{
Andy Liaw \email{andy_liaw@merck.com}.
}
