% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.rfCountData.R
\name{predict.rfCountData}
\alias{predict.rfCountData}
\title{predict method for random forest Count Data objects}
\usage{
\method{predict}{rfCountData}(object, newdata, offset, type = "response",
  norm.votes = TRUE, predict.all = FALSE, proximity = FALSE,
  nodes = FALSE, cutoff, ...)
}
\arguments{
\item{object}{an object of class \code{rfCountData}, as that created by the function \code{rf}}

\item{newdata}{a data frame or matrix containing new data.
(Note: If not given, the out-of-bag prediction in \code{object} is returned.}

\item{offset}{a vector with a length equal to the number of rows in \code{newdata}. Log of time of exposure.}

\item{type}{To be removed, or score scale vs. response scale ?}

\item{norm.votes}{not used}

\item{predict.all}{Should the predictions of all trees be kept?}

\item{proximity}{To be removed}

\item{nodes}{Should the terminal node indicators (an n by ntree matrix) be return?
If so, it is in the 'nodes' attribute of the returned object.}

\item{cutoff}{To be removed}

\item{...}{Currently not used}
}
\value{
{A vector of predicted values is returned.  \cr
If \code{predict.all=TRUE}, then the returned object is a list of two components: \code{aggregate}, 
which is the vector of predicted values by the forest, and \code{individual}, which
is a matrix where each column contains prediction by a tree in the forest.\cr
If \code{predict.all=TRUE}, then the \code{individual} component of thereturned object is a 
character matrix where each column contains the predicted class by a tree in the forest. \cr
If \code{nodes=TRUE}, the returned object has a 'nodes' attribute, which is an n by ntree matrix,
 each column containing the node number that the cases fall in for that tree.}
}
\description{
Prediction of test data using random forest.
}
\details{
predict.rfCountData
}
\references{
Breiman, L. (2001), \emph{Random Forests}, Machine Learning 45(1),5-32.
}
\seealso{
\code{\link{rfPoisson}}
}
\author{
Andy Liaw \email{andy_liaw@merck.com} and Matthew Wiener \email{matthew_wiener@merck.com}, 
 based on original Fortran code by  Leo Breiman and Adele Cutler.
}
\keyword{regression}
