% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfcv.R
\name{rfcv}
\alias{rfcv}
\title{rfcv}
\usage{
rfcv(trainx, trainy, cv.fold=5, scale="log", step=0.5, mtry=function(p) max(1, floor(sqrt(p))), recursive=FALSE, ...)
}
\arguments{
\item{trainx, }{matrix or data frame containing columns of predictor variables}

\item{trainoffset, }{vector of offset, must have length equal to the number of rows in \code{trainx}}

\item{trainy, }{vector of response, must have length equal to the number of rows in \code{trainx}}

\item{cv.fold, }{number of folds in the cross-validation}

\item{scale, }{if \code{"log"}, reduce a fixed proportion (\code{step}) of variables at each step, otherwise reduce step variables at a time}

\item{step, }{if \code{log=TRUE}, the fraction of variables to remove at each \code{step}, else remove this many variables at a time}

\item{mtry, }{a function of number of remaining predictor variables to use as the \code{mtry} parameter in the \code{rfPoisson} call}

\item{recursive, }{whether variable importance is (re-)assessed at each step of variable reduction}

\item{..., }{other arguments passed on to \code{rfPoisson}}
}
\description{
This function shows the cross-validated prediction performance of models with sequentially reduced number of predictors (ranked by variable importance) via a nested cross-validation procedure.
}
\references{
Svetnik, V., Liaw, A., Tong, C. and Wang, T., “Application of Breiman's Random Forest to Modeling Structure-Activity Relationships of Pharmaceutical Molecules”, MCS 2004, Roli, F. and Windeatt, T. (Eds.) pp. 334-343.
}
\author{
Andy Liaw
}
