% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesize.R
\name{treesize}
\alias{treesize}
\title{treesize}
\usage{
treesize(x, terminal = TRUE)
}
\arguments{
\item{x, }{an object of class \code{rfCountData}, which contains a \code{forest} component.}

\item{terminal, }{count terminal nodes only (\code{TRUE}) or all nodes (\code{FALSE}}
}
\value{
A vector containing number of nodes for the trees in the \code{rfCountData} object.
}
\description{
treesize
}
\note{
The \code{rfCountData} object must contain the \code{forest} component; i.e., created with \code{rfPoisson(...,keep.forest=TRUE)}.
}
\author{
Andy Liaw \email{andy_liaw@merck.com}
}
