% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varImpPlot.R
\name{varImpPlot}
\alias{varImpPlot}
\title{varImpPlot}
\usage{
varImpPlot(x, sort=TRUE, n.var=min(30, nrow(x$importance)),
type=NULL, class=NULL, scale=TRUE, 
main=deparse(substitute(x)), ...)
}
\arguments{
\item{x, }{An object of class \code{rfPoisson}.}

\item{sort, }{Should the variables be sorted in decreasing order of importance?}

\item{n.var, }{How many variables to show? (Ignored if \code{sort=FALSE}).}

\item{type, }{argument to be passed on to \code{\link{importance}}}

\item{class, }{argument to be passed on to \code{\link{importance}}}

\item{scale, }{argument to be passed on to \code{\link{importance}}}

\item{main, }{plot title,}

\item{..., }{Other graphical parameters to be passed on to \code{\link{dotchart}}.}
}
\value{
Invisibly, the importance of the variables that were plotted.
}
\description{
Dotchart of variable importance as measured by a Random Forest
}
\author{
Andy Liaw \email{andy_liaw@merck.com}.
}
